backend/requirements.txt
Flask==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.1
argon2-cffi==23.1.0